Fix: try to handle missing/non-existing crontab more gracefully
Fixes the following error in a chef-run:
DEBUG: LocalJumpError: easybib_crontab[gearmanworker] had an error: LocalJumpError: unexpected return

+++ easybib/libraries/cron.rb
@@ -12,10 +12,14 @@ module EasyBib
     def parse!
       regex = '([0-9/\-\*]+) +([0-9/\-\*]+) +([0-9/\-\*]+) +([0-9/\-\*]+) +([0-9/\-\*]+) +(.*)'
 
+      collection = []
+
+      if !::File.exists?(@file)
+        return collection
+      end
+
       crontabs = ::File.open(@file)
       cron_counter = 1
-
-      collection = []
 
       crontabs.each_line do |line|
         crontab = line.match(regex)
